Introduced Flyway and a 'dev' profile.

+++ src/main/java/rystad/playground/springboot/Sentence.java
@@ -7,11 +7,11 @@ import java.util.Random;
 @Entity(name = "sentences")
 public class Sentence {
     @Id
-    private Long id;
+    private Integer id;
     private String sentence;
 
     Sentence(final String sentence) {
-        this.id = new Random().nextLong();
+        this.id = new Random().nextInt();
         this.sentence = sentence;
     }
