+++ src/main/java/rystad/playground/springboot/SentenceDAO.java
@@ -2,5 +2,5 @@ package rystad.playground.springboot;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface SentenceDAO extends JpaRepository<Sentence, Long> {
+interface SentenceDAO extends JpaRepository<Sentence, Integer> {
 }
